requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
h11==0.14.0
idna==3.6
oauthlib==3.2.2
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
splitwise==3.0.0
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
